Noticeably improved speed of mobs with fixed point calculations
A scaler of 10 was used rather than 16 since it works better with the framerate and the current velocities. Otherwise the player alternates between two pixels while moving right

+++ source/mobs/ItemMob.cpp
@@ -55,7 +55,7 @@ void ItemMob::updateMob(WorldObject* world)
 	}
 	if (isBlockWalkThrough(world->blocks[int(x - sx / 2 + 1) / 16][int(y + sy / 2) / 16]) && isBlockWalkThrough(world->blocks[int(x + sx / 2) / 16][int(y + sy / 2) / 16]))
 	{
-		y += 16.0 * vy / double(FPS);
+		y += (16 * vy) / FPS;
 		vy += (18.0 / 60.0);
 	}
 	if (vy != 0 && !isBlockWalkThrough(world->blocks[int(x) / 16][int(y + sy / 2 - 1) / 16]))
